Make tests less fragile by using a wait after each URL get. Make tests faster by using a single driver per module.
