+++ tests_cfp__selenium/constants.py
@@ -1,0 +1,1 @@
+driver_wait_time = 2
