+++ tests_cfp__selenium/test_s_registered_user_can_login_submit_and_amend_proposals.py
@@ -5,28 +5,11 @@ from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as ecs
 
 from configuration import base_url
+from constants import driver_wait_time
 
 # NB PyCharm can't tell these are used as fixtures, but they are.
 # NB server is an session scope autouse fixture that no test needs direct access to.
-from fixtures import driver, server
-
-# NB The server is in "call open" state.
-
-
-@pytest.fixture
-def registrant():
-    return {
-        'email': 'a@b.c',
-        'passphrase': 'Passphrase1',
-        'cpassphrase': 'Passphrase1',
-        'name': 'User Name',
-        'phone': '+011234567890',
-        'country': 'India',
-        'state': 'TamilNadu',
-        'postal_code': '123456',
-        'town_city': 'Chennai',
-        'street_address': 'Chepauk',
-    }
+from fixtures import driver, server, registrant
 
 user_is_registered = False
 
@@ -34,7 +17,7 @@ user_is_registered = False
 def register_user(driver, registrant):
     if not user_is_registered:
         driver.get(base_url + 'register')
-        wait = WebDriverWait(driver, 4)
+        wait = WebDriverWait(driver, driver_wait_time)
         wait.until(ecs.text_to_be_present_in_element((By.CLASS_NAME, 'pagetitle'), ' – Register'))
         for key, value in registrant.items():
             driver.find_element_by_id(key).send_keys(value)
@@ -53,7 +36,7 @@ def register_user(driver, registrant):
 def test_user_can_successfully_login(driver, registrant):
     register_user(driver, registrant)
     driver.get(base_url + 'login')
-    wait = WebDriverWait(driver, 4)
+    wait = WebDriverWait(driver, driver_wait_time)
     wait.until(ecs.text_to_be_present_in_element((By.CLASS_NAME, 'pagetitle'), ' – Login'))
     driver.find_element_by_id('email').send_keys(registrant['email'])
     driver.find_element_by_id('passphrase').send_keys(registrant['passphrase'])
